Refactored Paging related code
The most common scenario of paging is that we’re dealing with 1 type of
object; however, there are certain scenarios where you might also want
to display a header (which requires a different type of object).
Instead of exposing all that nasty looking code in the fragment layer,
I refactored the current paging code to abstract away all the ugly
nuances of fetching that first page of data. Yes, the code is insanely
ugly, but it’s all abstracted behind high-level functionality (this
means the only time you have to touch the ugly code is if you’re
modifying it directly). I tested this for regressions by rerunning the
app on my phone, and seeing that paging still works for the
AccountFragment (it does).